+++ src/components/ButtonView/ButtonView.test.tsx
@@ -228,6 +228,260 @@ describe('ButtonView', () => {
     expect(persistentStorageMock.remove).toHaveBeenCalled();
   });
 
+  /**
+   * Reset
+   */
+  it('Should reset variable options with All option', () => {
+    const variable = {
+      ...deviceVariable,
+      options: deviceVariable.options.map((option) =>
+        option.value === 'device1'
+          ? {
+              ...option,
+              selected: true,
+            }
+          : option
+      ),
+      current: {
+        value: ['device1'],
+      },
+    };
+    jest.mocked(useRuntimeVariables).mockImplementationOnce(
+      () =>
+        ({
+          variable,
+        }) as any
+    );
+
+    render(
+      getComponent({
+        options: {
+          variable: 'device',
+          status: 'last',
+          persistent: true,
+          showResetButton: true,
+        } as any,
+      })
+    );
+
+    expect(selectors.item(false, 'device1')).toBeInTheDocument();
+    expect(selectors.resetVariable()).toBeInTheDocument();
+
+    /**
+     * Reset variable
+     */
+    fireEvent.click(selectors.resetVariable());
+
+    /**
+     * All should be selected
+     */
+    expect(selectVariableValues).toHaveBeenCalledWith({
+      panelEventBus: expect.anything(),
+      runtimeVariable: variable,
+      values: [ALL_VALUE_PARAMETER],
+    });
+  });
+
+  it('Should reset variable options for multi variable without All option', () => {
+    const variable = {
+      ...deviceVariable,
+      includeAll: false,
+      options: [
+        {
+          text: 'device1',
+          value: 'device1',
+          selected: false,
+        },
+        {
+          text: 'device2',
+          value: 'device2',
+          selected: false,
+        },
+        {
+          text: 'device3',
+          value: 'device3',
+          selected: true,
+        },
+        {
+          text: 'device4',
+          value: 'device4',
+          selected: false,
+        },
+      ],
+      current: {
+        value: ['device3'],
+      },
+    };
+    jest.mocked(useRuntimeVariables).mockImplementationOnce(
+      () =>
+        ({
+          variable,
+        }) as any
+    );
+
+    render(
+      getComponent({
+        options: {
+          variable: 'device',
+          status: 'last',
+          persistent: true,
+          showResetButton: true,
+        } as any,
+      })
+    );
+
+    expect(selectors.resetVariable()).toBeInTheDocument();
+
+    /**
+     * Reset variable
+     */
+    fireEvent.click(selectors.resetVariable());
+
+    /**
+     * device1 should be selected
+     */
+    expect(selectVariableValues).toHaveBeenCalledWith({
+      panelEventBus: expect.anything(),
+      runtimeVariable: variable,
+      values: ['device1'],
+    });
+  });
+
+  it('Should reset variable options for none multi variable', () => {
+    const variable = {
+      ...deviceVariable,
+      multi: false,
+      includeAll: false,
+      options: [
+        {
+          text: 'device1',
+          value: 'device1',
+          selected: false,
+        },
+        {
+          text: 'device2',
+          value: 'device2',
+          selected: false,
+        },
+        {
+          text: 'device3',
+          value: 'device3',
+          selected: true,
+        },
+        {
+          text: 'device4',
+          value: 'device4',
+          selected: false,
+        },
+      ],
+      current: {
+        value: 'device3',
+      },
+    };
+    jest.mocked(useRuntimeVariables).mockImplementationOnce(
+      () =>
+        ({
+          variable,
+        }) as any
+    );
+
+    render(
+      getComponent({
+        options: {
+          variable: 'device',
+          status: 'last',
+          persistent: true,
+          showResetButton: true,
+        } as any,
+      })
+    );
+
+    expect(selectors.resetVariable()).toBeInTheDocument();
+
+    /**
+     * Reset variable
+     */
+    fireEvent.click(selectors.resetVariable());
+
+    /**
+     * device1 should be selected
+     */
+    expect(selectVariableValues).toHaveBeenCalledWith({
+      panelEventBus: expect.anything(),
+      runtimeVariable: variable,
+      values: ['device1'],
+    });
+  });
+
+  /**
+   * Variable label
+   */
+  it('Should display variable label with label property', () => {
+    const variable = {
+      ...deviceVariable,
+      label: 'Test Label Variable',
+      options: deviceVariable.options.concat([
+        {
+          value: 'unknownDevice',
+          text: 'Unknown Device',
+          selected: true,
+        },
+      ]),
+    };
+    jest.mocked(useRuntimeVariables).mockImplementationOnce(
+      () =>
+        ({
+          variable,
+        }) as any
+    );
+
+    render(
+      getComponent({
+        options: {
+          variable: 'device',
+          status: 'last',
+          showLabel: true,
+        } as any,
+      })
+    );
+
+    expect(selectors.label()).toBeInTheDocument();
+    expect(selectors.label()).toHaveTextContent('Test Label Variable');
+  });
+
+  it('Should display variable label with name property', () => {
+    const variable = {
+      ...deviceVariable,
+      name: 'Test Name Variable',
+      options: deviceVariable.options.concat([
+        {
+          value: 'unknownDevice',
+          text: 'Unknown Device',
+          selected: true,
+        },
+      ]),
+    };
+    jest.mocked(useRuntimeVariables).mockImplementationOnce(
+      () =>
+        ({
+          variable,
+        }) as any
+    );
+
+    render(
+      getComponent({
+        options: {
+          variable: 'device',
+          status: 'last',
+          showLabel: true,
+        } as any,
+      })
+    );
+
+    expect(selectors.label()).toBeInTheDocument();
+    expect(selectors.label()).toHaveTextContent('Test Name Variable');
+  });
+
   it('Should work if no status color', () => {
     const variable = {
       ...deviceVariable,