+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Features / Enhancements
 
-- Add collapse/expand rows option on initial load (#159)
-- Update Selection Behavior using Ctrl key (#160)
+- Added collapse/expand rows option on initial load (#159)
+- Updated Selection Behavior using Ctrl key (#160)
+- Updated Grafana 11.0.0 and dependencies (#161)
+- Added Reset button for button view (#162)
 
 ## 3.0.0 (2024-05-09)